# public/openapi.yaml
openapi: 3.1.0
info:
  title: Aegis API
  version: 1.0.0
  summary: HTTP surface for the Aegis D2A social agent platform.
  description: |
    Aegis is a decentralized social agent platform that scores and curates
    content across Nostr, Farcaster, RSS, X, and arbitrary URLs using the
    V/C/L rubric (V=signal, C=context, L=slop) with a multi-tier scoring
    cascade (local LLMs -> BYOK Claude -> IC LLM -> server Claude ->
    heuristic fallback). This document describes the public HTTP endpoints
    exposed by the Next.js app at https://aegis-ai.xyz/api.
    D2A endpoints additionally implement the x402 payment protocol and
    CORS for cross-origin agent-to-agent consumption.
  license:
    name: MIT
    url: https://github.com/dwebxr/aegis/blob/main/LICENSE
  contact:
    name: Aegis on GitHub
    url: https://github.com/dwebxr/aegis
servers:
  - url: https://aegis-ai.xyz
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/d2a/score:
    get:
      tags: [D2A, Scoring]
      summary: Score a URL's content quality (x402-paid)
      description: |
        Extracts an HTTP(S) article and scores it with Claude using Aegis's
        V/C/L rubric. Production callers must provide an x402 v2
        `PAYMENT-SIGNATURE`; settlement occurs only after a successful 2xx
        handler response. The URL is sent as a query parameter and can appear
        in access logs. Identical URL/model results are cached for one hour.
      operationId: getD2AScore
      security:
        - x402: []
      parameters:
        - name: url
          in: query
          required: true
          description: HTTP(S) article URL to extract and score. Query order is preserved; fragments are ignored for caching.
          schema:
            type: string
            format: uri
            maxLength: 2048
      responses:
        '200':
          description: URL extraction and V/C/L scoring result.
          headers:
            PAYMENT-RESPONSE:
              $ref: '#/components/headers/PaymentResponse'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/D2AScoreResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          description: >
            x402 payment required, or settlement failed after scoring. An
            unpaid response carries `PAYMENT-REQUIRED`; a settlement response
            carries `PAYMENT-RESPONSE`.
          headers:
            PAYMENT-REQUIRED:
              $ref: '#/components/headers/PaymentRequired'
            PAYMENT-RESPONSE:
              $ref: '#/components/headers/PaymentResponse'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/X402SettlementFailureResponse'
        '422':
          description: The page contained no sufficiently meaningful article text.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '502':
          description: Article extraction, scoring, or payment facilitator failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: >
            Endpoint disabled/unconfigured, KV unavailable, daily budget
            exhausted, the same payment is already doing work, or another
            score for this URL is still in progress. In-progress responses
            use `Retry-After: 10`; markers expire naturally and are never
            explicitly deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    options:
      tags: [D2A]
      summary: CORS preflight
      operationId: d2aScoreOptions
      responses:
        '204':
          description: Preflight response with CORS headers.
...
components:
  securitySchemes:
    byokApiKey:
      type: apiKey
      in: header
      name: x-user-api-key
      description: |
        User's Anthropic API key ("Bring Your Own Key"). Must start with
        `sk-ant-`. When present, server budget and operator key are
        bypassed.
    pushToken:
      type: apiKey
      in: header
      name: token
      description: |
        HMAC-SHA256 push token obtained from `/api/push/token`. (Note: the
        token is passed in the JSON body, not a header, on /api/push/send;
        this security scheme documents the credential semantically.)
    x402:
      type: apiKey
      in: header
      name: PAYMENT-SIGNATURE
      description: |
        Base64 x402 v2 payment payload (see https://x402.org). Applies to
        `/api/d2a/briefing`, `/api/d2a/briefing/changes`, and
        `/api/d2a/score` when their payment gates are enabled. Briefing
        preview routes can bypass payment when their free tier is enabled;
        score requires payment in production.
    openPayX402:
      type: apiKey
      in: header
      name: X-PAYMENT
      description: >
        OpenPay-flavored x402 v1 payment payload for
        `/api/d2a/briefing-jpyc` on Polygon.
  headers:
    PaymentRequired:
      description: Base64-encoded x402 v2 PaymentRequired object.
      schema:
        type: string
    PaymentResponse:
      description: Base64-encoded x402 v2 settlement result.
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request (missing/malformed input).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Too many requests — see `Retry-After` header.
      headers:
        Retry-After:
          schema:
            type: integer
            description: Seconds until the rate-limit window resets.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        reason:
          type: string
    X402SettlementFailureResponse:
      type: object
      description: >
        x402 v2 error body. The library currently emits an empty JSON object
        for unpaid and settlement-failure responses; PAYMENT-REQUIRED or
        PAYMENT-RESPONSE is authoritative.
      additionalProperties: true
    D2AScoreResponse:
      type: object
      required: [url, title, source, scoredAt, engine, model, cached, score]
      properties:
        url:
          type: string
          format: uri
        title:
          type: string
        source:
          type: string
        author:
          type: string
        publishedDate:
          type: string
        scoredAt:
          type: string
          format: date-time
        engine:
          type: string
          enum: [claude]
        model:
          type: string
        cached:
          type: boolean
        score:
          type: object
          required: [vSignal, cContext, lSlop, originality, insight, credibility, composite, verdict, reason, topics]
          properties:
            vSignal:
              type: number
              minimum: 0
              maximum: 10
            cContext:
              type: number
              minimum: 0
              maximum: 10
            lSlop:
              type: number
              minimum: 0
              maximum: 10
            originality:
              type: number
              minimum: 0
              maximum: 10
            insight:
              type: number
              minimum: 0
              maximum: 10
            credibility:
              type: number
              minimum: 0
              maximum: 10
            composite:
              type: number
              minimum: 0
              maximum: 10
            verdict:
              $ref: '#/components/schemas/Verdict'
            reason:
              type: string
            topics:
              type: array
              items:
                type: string
    Verdict:
      type: string
      enum: [quality, slop]
...